- updated instructions for testcase - updated testcase in the bookmarklet
