+++ oabutton/oabutton/apps/bookmarklet/models.py
@@ -19,7 +19,7 @@ class Event(models.Model):
 
     def set_coords(self, coords):
         """Expect a comma separated string of float """
-        self.coords_lat, self.coords_lng = coords['lat'], coords['lng']
+        self.coords_lat, self.coords_lng = str(coords['lat']), str(coords['lng'])
 
     def get_coords(self):
         """Return a comma separated string of floats"""
